move MODIS fetch functions to Fetcher (not tested)
