+++ Project.toml
@@ -1,16 +1,13 @@
 name = "GriddingMachine"
 uuid = "f20cf718-bf4d-4727-bc8f-485b1f283ac6"
 authors = ["Yujie Wang"]
-version = "0.2.2"
+version = "0.2.3"
 
 [deps]
 Artifacts = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
 Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
-Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
@@ -18,17 +15,13 @@ PkgUtility = "0d262f2c-28e9-492c-8e19-d7a5c4f11611"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
 Conda = "1.5.2"
-DataFrames = "1.2.2"
 DocStringExtensions = "0.8.5"
-Glob = "1.3.0"
 HTTP = "0.9.14"
 JSON = "0.21.2"
 PkgUtility = "0.1.13"
 ProgressMeter = "1.7.1"
 PyCall = "1.92.3"
-UnPack = "1.0.2"
 julia = "1.6"
