+++ src/GriddingMachine.jl
@@ -7,7 +7,6 @@ include("Collector.jl")
 include("Fetcher.jl")
 include("Indexer.jl")
 include("Requestor.jl")
-include("GriddingMachineOld.jl")
 
 
 end # module
