+++ src/deprecated/Gridding/compile.jl
@@ -1,3 +1,4 @@
+#=
 ###############################################################################
 #
 # Compile data to nc files
@@ -224,3 +225,4 @@ function compile_RAW!(
 
     return nothing
 end
+=#
