+++ src/deprecated/Gridding/load.jl
@@ -1,3 +1,4 @@
+#=
 ###############################################################################
 #
 # Load MODIS grid information
@@ -11,7 +12,6 @@ Prepare parameters (file name and etc) to work on, given
 """
 function load_MODIS(dt::AbstractMODIS500m{FT}, h::Int, v::Int) where {FT<:AbstractFloat}
     # read MODIS gridding info
-    predownload_artifact!("MODIS_500m_grid", ARTIFACTs_TOML);
     _file = artifact"MODIS_500m_grid" * "/MODIS_500m_grid.nc";
 
     _lat_mat = read_nc(FT, _file, "latitude" , h, v);
@@ -25,7 +25,6 @@ end
 
 function load_MODIS(dt::AbstractMODIS1km{FT}, h::Int, v::Int) where {FT<:AbstractFloat}
     # read MODIS gridding info
-    predownload_artifact!("MODIS_1km_grid", ARTIFACTs_TOML);
     _file = artifact"MODIS_1km_grid" * "/MODIS_1km_grid.nc";
 
     _lat_mat = read_nc(FT, _file, "latitude" , h, v);
@@ -33,3 +32,4 @@ function load_MODIS(dt::AbstractMODIS1km{FT}, h::Int, v::Int) where {FT<:Abstrac
 
     return _lat_mat,_lon_mat
 end
+=#
