+++ src/deprecated/Gridding/shortcuts.jl
@@ -1,3 +1,4 @@
+#=
 ###############################################################################
 #
 # Shortcut to do all steps within one function
@@ -36,3 +37,4 @@ function process_RAW!(
 
     return nothing
 end
+=#
